[[action:merge_dev]] avatar image reverted

+++ accounts/signals.py
@@ -1,5 +1,6 @@
 import os
 
+from django.core.files.storage import FileSystemStorage
 from django.db.models.signals import post_delete
 from django.dispatch import receiver
 
@@ -18,14 +19,16 @@ def delete_user_avatar(sender, instance: User, **kwargs) -> None:
         return
 
     # Dosyayı sil (storage üzerinden, OverWriteAvatarStorage dahil)
-    avatar_path = avatar.path  # type: ignore
+    storage = avatar.storage
+    avatar_name = avatar.name
     avatar.delete(save=False)  # type: ignore
 
-    # Klasörü boşsa klasörü de kaldır
-    folder = os.path.dirname(avatar_path)
-    try:
-        if os.path.isdir(folder) and not os.listdir(folder):
-            os.rmdir(folder)
-    except OSError:
-        # Herhangi bir izin / race condition hatasında sessizce geç
-        pass
+    # Klasörü boşsa klasörü de kaldır (yalnızca yerel dosya sistemi)
+    if isinstance(storage, FileSystemStorage):
+        try:
+            folder = os.path.dirname(storage.path(avatar_name))
+            if os.path.isdir(folder) and not os.listdir(folder):
+                os.rmdir(folder)
+        except (OSError, NotImplementedError):
+            # Herhangi bir izin / race condition hatasında sessizce geç
+            pass
